Remove 'Network: true' flag from plugin metadata.
Don't force network-activate the plugin on multisite installs.  This
decision should be left up to the network admin.

However, if you activate only on the main BP site, activity hashtag
recording will not work when a sub-site publishes a new blog post or posts
a new comment containing a hashtag.

See #7.